zip deletion metric fix

+++ bin/helpers/fileHelpers.js
@@ -27,12 +27,15 @@ exports.fileExists = function (filePath, cb) {
 
 exports.deleteZip = () => {
   try {
+    if (!fs.existsSync(config.fileName)) {
+      return 0;
+    }
     fs.unlinkSync(config.fileName);
     logger.info(Constants.userMessages.ZIP_DELETED);
     return 0;
   } catch (err) {
     logger.info(Constants.userMessages.ZIP_DELETE_FAILED);
-    return 1;
+    throw err;
   }
 };
